requirements.txt
dateparser==1.1.1
Flask==2.0.0
Flask-Cors==3.0.10
Jinja2==3.0.0
peewee==3.15.3
python-dateutil==2.8.2
gunicorn==20.1.0
gevent==21.12.0
PyYAML>=6.0
apscheduler>=3.9.1
PyJWT>=2.5.0
bcrypt>=4.0.0